Rename gateway to hub
