+++ tests/components/deconz/snapshots/test_hub.ambr
