+++ tests/components/deconz/test_hub.py
